+++ fugitive-darkness-models/src/main/java/com/zer0s2m/fugitivedarkness/models/GitJobModel.java
@@ -91,4 +91,8 @@ public class GitJobModel extends Job<String> implements FieldObject {
         return nextRunAt;
     }
 
+    public boolean getIsGitRepoIsLocal() {
+        return gitRepoIsLocal;
+    }
+
 }