+++ fugitive-darkness-plugin/fugitive-darkness-plugin-vertx-job/src/main/java/com/zer0s2m/fugitivedarkness/plugin/vertx/job/impl/ListenerGitJobPermanent.java
@@ -5,7 +5,9 @@ import com.zer0s2m.fugitivedarkness.models.GitJobModel;
 import com.zer0s2m.fugitivedarkness.plugin.job.GitTypeJob;
 import com.zer0s2m.fugitivedarkness.plugin.job.JobException;
 import com.zer0s2m.fugitivedarkness.plugin.job.JobManager;
+import com.zer0s2m.fugitivedarkness.plugin.job.Rule;
 import com.zer0s2m.fugitivedarkness.plugin.job.impl.JobManagerImpl;
+import com.zer0s2m.fugitivedarkness.plugin.job.rule.RuleIsLocalGitRepo;
 import com.zer0s2m.fugitivedarkness.plugin.job.support.CronExpression;
 import com.zer0s2m.fugitivedarkness.plugin.vertx.job.Listener;
 import com.zer0s2m.fugitivedarkness.repository.GitJobRepository;
@@ -29,6 +31,12 @@ public class ListenerGitJobPermanent implements Listener {
     final private FieldBuilder<Row> fieldBuilder = new FieldBuilderRow();
 
     final private JobManager jobManager = new JobManagerImpl();
+
+    final private Rule ruleIsLocalGitRepo = new RuleIsLocalGitRepo();
+
+    {
+        ruleIsLocalGitRepo.setTypeTask(GitTypeJob.PERMANENT);
+    }
 
     /**
      * Run a listening scan from different sources.
@@ -58,6 +66,9 @@ public class ListenerGitJobPermanent implements Listener {
                                 isUpdateRow = true;
 
                                 try {
+                                    jobManager.clearRule();
+                                    jobManager.addRule(ruleIsLocalGitRepo);
+
                                     jobManager.call(
                                             GitTypeJob.PERMANENT,
                                             ListenerGitJobUtils.collectProperties(gitJob));